move update_filter_values to filter class
